pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "autogen-agentchat>=0.7.5",
    "autogen-core>=0.7.5",
    "autogen-ext>=0.7.5",
    "fastapi>=0.135.3",
    "jinja2>=3.1.6",
    "openai>=2.31.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.44.0",
    "websockets>=16.0",
]
